+++ Tests/Tests.Shared/TestHelpers.cs
@@ -143,14 +143,6 @@ namespace Tests
             }
         }
 
-        public static void ReliesOnEncryption()
-        {
-            if (IsWindows)
-            {
-                Assert.Ignore("This test relies on encryption which is not enabled in this build.");
-            }
-        }
-
         public static RealmInteger<T>[] ToInteger<T>(this T[] values)
             where T : struct, IComparable<T>, IFormattable
         {